+++ openstack/table_openstack_server.go
@@ -3,7 +3,6 @@ package openstack
 import (
 	"context"
 
-	"github.com/gophercloud/gophercloud"
 	"github.com/gophercloud/gophercloud/openstack"
 	"github.com/gophercloud/gophercloud/openstack/compute/v2/servers"
 	"github.com/turbot/steampipe-plugin-sdk/v5/grpc/proto"
@@ -61,7 +60,8 @@ func listServer(ctx context.Context, d *plugin.QueryData, _ *plugin.HydrateData)
 	}
 
 	// get compute client from provider
-	computeClient, err := openstack.NewComputeV2(provider, gophercloud.EndpointOpts{})
+	endpointOpts := getEndpointOpts(d)
+	computeClient, err := openstack.NewComputeV2(provider, endpointOpts)
 
 	if err != nil {
 		logger.Error("openstack_server.listServer", "connection_error", err)
@@ -96,7 +96,8 @@ func getServer(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) 
 	}
 
 	// get compute client from provider
-	computeClient, err := openstack.NewComputeV2(provider, gophercloud.EndpointOpts{})
+	endpointOpts := getEndpointOpts(d)
+	computeClient, err := openstack.NewComputeV2(provider, endpointOpts)
 
 	// get server
 	server, err := servers.Get(computeClient, id).Extract()